+++ core/block_validator.go
@@ -41,12 +41,12 @@ import (
 // BlockValidator implements validator.
 type BlockValidator struct {
 	config *params.ChainConfig     // Chain configuration options
-	bc     *BlockChain             // Canonical block chain
+	bc     BlockChain              // Canonical blockchain
 	engine consensus_engine.Engine // Consensus engine used for validating
 }
 
 // NewBlockValidator returns a new block validator which is safe for re-use
-func NewBlockValidator(config *params.ChainConfig, blockchain *BlockChain, engine consensus_engine.Engine) *BlockValidator {
+func NewBlockValidator(config *params.ChainConfig, blockchain BlockChain, engine consensus_engine.Engine) *BlockValidator {
 	validator := &BlockValidator{
 		config: config,
 		engine: engine,
