+++ node/node_handler.go
@@ -280,7 +280,7 @@ func (node *Node) BroadcastCrosslinkHeartbeatSignalFromBeaconToShards() { // lea
 }
 
 // getCrosslinkHeadersForShards get headers required for crosslink creation.
-func getCrosslinkHeadersForShards(beacon *core.BlockChain, shardChain *core.BlockChain, curBlock *types.Block, shardID uint32, latestSentCrosslink *uint64) ([]*block.Header, error) {
+func getCrosslinkHeadersForShards(beacon core.BlockChain, shardChain core.BlockChain, curBlock *types.Block, shardID uint32, latestSentCrosslink *uint64) ([]*block.Header, error) {
 	var headers []*block.Header
 	lastLink, err := beacon.ReadShardLastCrossLink(shardID)
 	var latestBlockNum uint64
